review-service/app/database.py
# Conexão com MongoDB (assíncrona com motor)

import motor.motor_asyncio
import os
from dotenv import load_dotenv

load_dotenv()

MONGO_URL = os.getenv("MONGO_URL", "mongodb://localhost:27017/reviews")
client = motor.motor_asyncio.AsyncIOMotorClient(MONGO_URL)
db = client.get_default_database()
reviews_collection = db.get_collection("reviews")
