review-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0
